requirements.txt
PyPDF2
pymongo
openai
python-dotenv
python-dotenv
pydantic
requests
numpy
schedule
